backend/requirements.txt
python-dotenv
google-generativeai
Flask
flask-cors
